requirements.txt
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.2.7
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.3
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.2
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.1
drf-yasg==1.21.7
eventlet==0.33.3
exceptiongroup==1.1.2
executing==1.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.19.0
kombu==5.3.1
matplotlib-inline==0.1.6
packaging==23.1
parso
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6